test: split treemap
