+++ .storybook/config.ts
@@ -51,6 +51,7 @@ function loadStories() {
   require('../stories/annotations.tsx');
   require('../stories/scales.tsx');
   require('../stories/sunburst.tsx');
+  require('../stories/treemap.tsx');
 }
 
 configure(loadStories, module);
